pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ultrafastrss"
dynamic = ["version"]
description = 'Async Simple and fast RSS parser module.'
readme = "README.md"
requires-python = ">=3.8"
license = "GPL-3.0-or-later"
keywords = []
authors = [
  { name = "Maikel Mardjan", email = "mike@bm-support.org" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]

dependencies = ["Brotli", 
"python-dateutil",
]

[project.urls]
Documentation = "https://github.com/Maikel Mardjan/ultrafastrss#readme"
Issues = "https://github.com/Maikel Mardjan/ultrafastrss/issues"
Source = "https://github.com/Maikel Mardjan/ultrafastrss"

[tool.hatch.version]
path = "src/ultrafastrss/__about__.py"

[tool.hatch.envs.types]
extra-dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/ultrafastrss tests}"

[tool.coverage.run]
source_pkgs = ["ultrafastrss", "tests"]
branch = true
parallel = true
omit = [
  "src/ultrafastrss/__about__.py",
]

[tool.coverage.paths]
ultrafastrss = ["src/ultrafastrss", "*/ultrafastrss/src/ultrafastrss"]
tests = ["tests", "*/ultrafastrss/tests", "*/ultrafastrss/tests/resources"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]

